refactor: consolidate and enhance parsing functions for non-negative and positive integers
